fix dtypes

+++ efficientdet/anchors.py
@@ -334,13 +334,13 @@ def _generate_detections_tf(cls_outputs,
 
   image_size = utils.parse_image_size(image_size)
   detections = tf.stack([
-      tf.cast(tf.tile(image_id, tf.shape(top_detection_idx)), tf.float32),
+      tf.cast(tf.tile(image_id, tf.shape(top_detection_idx)), image_scale.dtype),
       tf.clip_by_value(boxes[:, 0], 0, image_size[0]) * image_scale,
       tf.clip_by_value(boxes[:, 1], 0, image_size[1]) * image_scale,
       tf.clip_by_value(boxes[:, 2], 0, image_size[0]) * image_scale,
       tf.clip_by_value(boxes[:, 3], 0, image_size[1]) * image_scale,
       scores,
-      tf.cast(tf.gather(classes, top_detection_idx) + 1, tf.float32)
+      tf.cast(tf.gather(classes, top_detection_idx) + 1, image_scale.dtype)
   ], axis=1)
   return detections
